+++ src/main/java/org/mockito/internal/handler/NotifiedMethodInvocationReport.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2007 Mockito contributors
  * This program is made available under the terms of the MIT License.
  */
-package org.mockito.internal.listeners;
+package org.mockito.internal.handler;
 
 import org.mockito.invocation.DescribedInvocation;
 import org.mockito.invocation.Invocation;
