--- src/main/java/org/mockito/internal/listeners/MockingProgressListener.java
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) 2007 Mockito contributors
- * This program is made available under the terms of the MIT License.
- */
-package org.mockito.internal.listeners;
-
-public interface MockingProgressListener {
-}
